+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2025-10-14
+
+## Changed
+
+- The search box now supports a cursor for better text editing in
+[app.rs](src/ui/app.rs)
+- Made the search box input prompt constant in [app.rs](src/ui/app.rs)
+
 ## [0.7.0] - 2025-10-10
 
 ## Added